Fix Remove Chat multiple calls and refactor Browser Utils
