+++ src/utils/browser-utils.ts
@@ -1,0 +1,11 @@
+export const isLastSession = () => {
+  const sessions = localStorage.getItem("sessions");
+  return sessions && parseInt(sessions) === 1;
+}
+
+export const wasPageReloaded = () => {
+  return window.performance
+    .getEntriesByType('navigation')
+    .map((nav) => (nav as PerformanceNavigationTiming).type)
+    .includes('reload');
+}
